Issue 233: Optionally rename the _type when there is a mapping mismatch
It may happen that ES rejects an event because a property of your event
is typed differently than it is in the ES mapping. This is expected
behaviour.

The current behaviour is to NOT retry and log the event. This commit
allows one to overwrite and make the plugin change the type to a new
semi-randomised name (with a new matching mapping, of course) and retry
to submit it.